+++ assets/jetbrains/src/main/kotlin/com/dream/DreamCompletionContributor.kt
@@ -1,0 +1,25 @@
+package com.dream
+
+import com.intellij.codeInsight.completion.*
+import com.intellij.codeInsight.lookup.LookupElementBuilder
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.util.ProcessingContext
+
+class DreamCompletionContributor : CompletionContributor() {
+    init {
+        extend(
+            CompletionType.BASIC,
+            PlatformPatterns.psiElement(),
+            object : CompletionProvider<CompletionParameters>() {
+                override fun addCompletions(
+                    parameters: CompletionParameters,
+                    context: ProcessingContext,
+                    result: CompletionResultSet
+                ) {
+                    result.addElement(LookupElementBuilder.create("Console.Write"))
+                    result.addElement(LookupElementBuilder.create("Console.WriteLine"))
+                }
+            }
+        )
+    }
+}